Potential fix
I may have found the issue. Seems to have resolved most of my problems.
I referred to this:
http://stackoverflow.com/questions/35675855/android-studio-floatingactio
nbutton-error
It seemed to be a common issue among people. If we need/want to change
this later, so be it. Right now, leave it alone while it’s in this
branch